[BetterPhpDocParser] Move PhpDocFromTypeDeclarationDecorator from DowngradePhp71 to BetterPhpDocParser
