+++ packages/BetterPhpDocParser/PhpDocParser/PhpDocFromTypeDeclarationDecorator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\DowngradePhp71\TypeDeclaration;
+namespace Rector\BetterPhpDocParser\PhpDocParser;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\Closure;
